requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.4
django-allauth==0.62.1
django-bootstrap4==24.3
django-crispy-bulma==0.2
django-crispy-forms==2.1
djangorestframework==3.15.1
idna==3.7
isodate==0.6.1
lxml==5.2.1
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.0
python3-openid==3.2.0
python3-saml==1.16.0
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
urllib3==2.2.1
xmlsec==1.3.14
